test: getAllTournamentList
